requirements.txt
markdown2>=2.4.10
click>=8.1.7
pathlib>=1.0.1
PyYAML>=6.0.1
python-frontmatter>=1.0.0
rich>=13.7.0
PyGithub>=2.1.1
requests>=2.31.0
python-dotenv>=1.0.0
questionary>=1.10.0
markdown>=3.4.3
beautifulsoup4>=4.9.3
selenium>=4.0.0
playwright>=1.20.0
httpx>=0.24.0
html2text>=2020.1.16
2captcha-python>=1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
lxml==4.9.3
html5lib==1.1
chardet==5.2.0
urllib3==2.1.0
certifi==2023.11.17
idna==3.6
multidict==6.0.4
yarl==1.9.4
frozenlist==1.4.0
attrs==23.1.0
exceptiongroup==1.2.0
packaging==23.2
pluggy==1.3.0
iniconfig==2.0.0
typing_extensions==4.8.0
async-timeout==4.0.3
aiosignal==1.3.1
humanize>=4.0.0
  